requirements.txt
absl-py==0.7.1
appdirs==1.4.4
ase==3.20.1
astor==0.7.1
attrs==21.2.0
backcall==0.1.0
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configparser==5.0.1
cycler==0.10.0
Cython==0.29.21
dataclasses==0.7
decorator==4.4.0
docker-pycreds==0.4.0
future==0.18.2
gast==0.2.2
gitdb==4.0.5
GitPython==3.1.13
google-pasta==0.1.7
googledrivedownloader==0.4
grpcio==1.20.1
h5py==2.9.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.5.0
iniconfig==1.1.1
install==1.3.3
ipython==7.8.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.15.1
Jinja2==2.11.2
joblib==0.16.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
llvmlite==0.34.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
mypy-extensions==0.4.3
networkx==2.4
numba==0.51.1
numpy==1.16.5
olefile==0.46
opencv-python==4.1.1.26
packaging==20.9
pandas==0.23.4
parso==0.5.1
pathlib==1.0.1
pathspec==0.8.0
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pluggy==0.13.1
plyfile==0.7.2
promise==2.3
prompt-toolkit==2.0.10
protobuf==3.15.1
psutil==5.8.0
ptyprocess==0.6.0
py==1.10.0
pycocotools==2.0.2
Pygments==2.4.2
pyparsing==2.4.2
pytest==6.2.4
python-dateutil==2.8.0
pytorch-nlp==0.5.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
rdflib==5.0.0
regex==2020.7.14
requests==2.24.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.1.0
sentry-sdk==0.20.3
shortuuid==1.0.1
six==1.15.0
smmap==3.0.5
subprocess32==3.5.4
tensorboard==1.15.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tifffile==2020.8.25
toml==0.10.1
torch==1.5.1
torch-cluster==1.5.7
torch-geometric==1.6.0
torch-scatter==2.0.5
torch-sparse==0.6.7
torch-spline-conv==1.2.0
torchsummary==1.5.1
torchvision==0.6.0
tqdm==4.36.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.10.0.0
urllib3==1.25.10
wandb==0.10.19
watchdog==0.10.3
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==3.4.1
